requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
idna==2.9
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
urllib3==1.25.9
validators==0.14.3
